+++ src/xrEngine/editor_environment_ambients_effect_id.cpp
@@ -41,7 +41,7 @@ LPCSTR const* effect_id::collection					()
 
 u32 effect_id::collection_size						()
 {
-	return							(m_manager.effects_ids().size());
+	return							((u32)m_manager.effects_ids().size());
 }
 
 void effect_id::fill								(editor::property_holder_collection* collection)
