+++ src/xrGame/agent_enemy_manager.cpp
@@ -237,7 +237,7 @@ void CAgentEnemyManager::assign_enemies			()
 		(*I).m_probability				*= 1.f - best; 
 
 		// recovering sort order
-		for (u32 i_=0, n = m_enemies.size() - 1; i_<n; ++i_)
+		for (u32 i_=0, n = (u32)m_enemies.size() - 1; i_<n; ++i_)
 			if (m_enemies[i_ + 1] < m_enemies[i_])
 				std::swap				(m_enemies[i_],m_enemies[i_ + 1]);
 			else
@@ -422,7 +422,7 @@ void CAgentEnemyManager::assign_wounded			()
 	VERIFY					(enemy_mask == object().member().combat_mask());
 #endif // DEBUG
 
-	u32						previous_wounded_count = m_wounded.size();
+	u32						previous_wounded_count = (u32)m_wounded.size();
 	WOUNDED_ENEMY			*previous_wounded = (WOUNDED_ENEMY*)_alloca(previous_wounded_count*sizeof(WOUNDED_ENEMY));
 	std::copy				(m_wounded.begin(),m_wounded.end(),previous_wounded);
 	m_wounded.clear			();
